backend/requirements.txt
# Django Framework
Django>=5.2.0
djangorestframework>=3.16.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
joblib>=1.3.0

# Data Processing
pandas>=2.0.0

# Optional: Logging and Monitoring
python-json-logger>=2.0.0
